Update Ext_T_AutoFireHelper.uc
I think this will accomplish the goal you outlined in the discussion, but I'm not sure, mainly because I don't know where to find where the object PawnOwner is defined - I'm imagining it should have a BasePerk associated with it, but I could be completely off base. Sorry if this is adding more workload than it's alleviating.